refactor: extract verifyargs
